Updated seed.js and moved movieSeedData to seperate file
