+++ src/server/db/seed.js
@@ -1,6 +1,10 @@
 const db = require('./client');
 const { createUser } = require('./users');
 const { createComment } = require('./comments')
+const moviesSeedData = require('./moviesSeedData'); // Adjust the path as necessary
+
+// ... rest of your seed logic ...
+
 
 const users = [
   {
@@ -42,88 +46,6 @@ const comments = [
     review_id: 2,
     user_id: 1,
   },
-];
-
-const moviesSeedData = [
-  {
-      title: "Inception",
-      description: "A thief who steals corporate secrets through the use of dream-sharing technology...",
-      genre: "Sci-Fi",
-      releaseYear: 2010,
-      rating: 8.8
-  },
-  {
-    title: "The Shawshank Redemption",
-    description: "Two imprisoned men bond over a number of years, finding solace and eventual redemption through acts of common decency.",
-    genre: "Drama",
-    releaseYear: 1994,
-    rating: 9.3
-},
-{
-    title: "The Godfather",
-    description: "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son.",
-    genre: "Crime, Drama",
-    releaseYear: 1972,
-    rating: 9.2
-},
-{
-    title: "The Dark Knight",
-    description: "When the menace known as the Joker wreaks havoc and chaos on the people of Gotham, Batman must accept one of the greatest psychological and physical tests of his ability to fight injustice.",
-    genre: "Action, Crime, Drama",
-    releaseYear: 2008,
-    rating: 9.0
-},
-{
-    title: "Pulp Fiction",
-    description: "The lives of two mob hitmen, a boxer, a gangster's wife, and a pair of diner bandits intertwine in four tales of violence and redemption.",
-    genre: "Crime, Drama",
-    releaseYear: 1994,
-    rating: 8.9
-},
-{
-    title: "Fight Club",
-    description: "An insomniac office worker and a devil-may-care soapmaker form an underground fight club that evolves into much more.",
-    genre: "Drama",
-    releaseYear: 1999,
-    rating: 8.8
-},
-{
-    title: "Forrest Gump",
-    description: "The presidencies of Kennedy and Johnson, the Vietnam War, the Watergate scandal and other historical events unfold from the perspective of an Alabama man with an IQ of 75, whose only desire is to be reunited with his childhood sweetheart.",
-    genre: "Drama, Romance",
-    releaseYear: 1994,
-    rating: 8.8
-},
-{
-    title: "Inglourious Basterds",
-    description: "In Nazi-occupied France during World War II, a plan to assassinate Nazi leaders by a group of Jewish U.S. soldiers coincides with a theatre owner's vengeful plans for the same.",
-    genre: "Adventure, Drama, War",
-    releaseYear: 2009,
-    rating: 8.3
-},
-{
-    title: "The Matrix",
-    description: "When a beautiful stranger leads computer hacker Neo to a forbidding underworld, he discovers the shocking truth--the life he knows is the elaborate deception of an evil cyber-intelligence.",
-    genre: "Action, Sci-Fi",
-    releaseYear: 1999,
-    rating: 8.7
-},
-{
-    title: "Se7en",
-    description: "Two detectives, a rookie and a veteran, hunt a serial killer who uses the seven deadly sins as his motives.",
-    genre: "Crime, Drama, Mystery",
-    releaseYear: 1995,
-    rating: 8.6
-},
-{
-    title: "Interstellar",
-    description: "A team of explorers travel through a wormhole in space in an attempt to ensure humanity's survival.",
-    genre: "Adventure, Drama, Sci-Fi",
-    releaseYear: 2014,
-    rating: 8.6
-},
-
-  // ... more movie data ...
 ];
 
 const reviewsSeedData = [
@@ -249,8 +171,8 @@ async function insertMovies() {
     for (const movie of moviesSeedData) {
       const { title, description, genre, releaseYear, rating } = movie;
       const result = await db.query(
-        `INSERT INTO movies (title, description, genre, release_year, rating) VALUES ($1, $2, $3, $4, $5) RETURNING *;`,
-        [title, description, genre, releaseYear, rating]
+        `INSERT INTO movies (title, description, genre, release_year, rating, created_at) VALUES ($1, $2, $3, $4, $5, $6) RETURNING *;`,
+        [title, description, genre, releaseYear, rating, new Date()]
       );
       console.log(result.rows[0]); 
     }
